swap new server and EBC scripts

+++ scripts/06.first_level/generate_eventfiles.sh
@@ -4,7 +4,7 @@
 # GENERATE THE EVENTS FILES USED TO IDENTIFY EVENTS OF INTEREST IN THE FUNCTIONAL DATA
 # 
 # Each functional run should have a corresponding events file 
-# The name of the functional run is indicated in the events file name (e.g., sub-001_ses-01_task-pixar_run-01_events.tsv)
+# The name of the functional run is indicated in the events file name (e.g., sub-001_task-KMVPAst_run-01_events.tsv)
 # In order for event files to be generated, there needs to be pre-existing events.tsv files in the project data directory
 #
 # More information on these files: 
@@ -19,11 +19,11 @@ Usage() {
 	echo "./generate_eventfiles.sh <subject-run list> <task>"
 	echo
 	echo "Example:"
-	echo "./generate_eventfiles.sh TEBC-5y_subjs.txt pixar"
+	echo "./generate_eventfiles.sh KMVPA_subjs.txt KMVPA"
 	echo 
-	echo "TEBC-5y_subjs.txt is a subject-run file containing the participants and run info to generate the events.tsv files for:"
-	echo "8010 1,2"
-	echo "8011 1,2"
+	echo "KMVPA_subjs.txt is a subject-run file containing the participants and run info to generate the events.tsv files for:"
+	echo "001 1,2"
+	echo "002 2"
 	echo "..."
 	echo
 	echo
@@ -34,30 +34,20 @@ Usage() {
 [ "$1" = "" ] | [ "$2" = "" ] && Usage
 
 # indicate whether session folders are used (always 'yes' for EBC data)
-sessions='yes'
+sessions='no'
 
 # define task from script call
 task=$2
 
-# extract sample from list of subjects filename (i.e., are these pilot or HV subjs)
-sample=` basename $1 | cut -d '-' -f 3 | cut -d '.' -f 1 `
+# extract study name from list of subjects filename
+study=` basename $1 | cut -d '_' -f 1 `
 
-# define data directories depending on sample information
-if [[ ${sample} == 'pilot' ]]
-then
-	bidsDir="/data/EBC/preprocessedData/TEBC-5y/BIDs_data/pilot"
-	derivDir="/data/EBC/preprocessedData/TEBC-5y/derivatives/pilot"
-elif [[ ${sample} == 'HV' ]]
-then
-	bidsDir="/data/EBC/preprocessedData/TEBC-5y-adultpilot/BIDs_data"
-	derivDir="/data/EBC/preprocessedData/TEBC-5y-adultpilot/derivatives"
-else
-	bidsDir="/data/EBC/preprocessedData/TEBC-5y/BIDs_data"
-	derivDir="/data/EBC/preprocessedData/TEBC-5y/derivatives"
-fi
+# define data directories depending on study information
+bidsDir="/RichardsonLab/preprocessedData/${study}"
+derivDir="${bidsDir}/derivatives"
 
 # print confirmation of sample and directory
-echo 'Generating events.tsv files for' ${sample} 'data in' ${derivDir}
+echo 'Generating events.tsv files for' ${study} 'data in' ${derivDir}
 
 # define project directory
 projDir=`cat ../../PATHS.txt`
@@ -72,17 +62,17 @@ do
 	# define subject derivatives directory depending on whether data are organized in session folders
 	if [[ ${sessions} == 'yes' ]]
 	then
-		subDir="${derivDir}/sub-${sub}/ses-01/func"
-		tsv_prefix="sub-${sub}_ses-01_task-${task}"
+		subDir="${derivDir}/${sub}/ses-01/func"
+		tsv_prefix="${sub}_ses-01_task-${task}"
 	else
-		subDir="${derivDir}/sub-${sub}/func"
-		tsv_prefix="sub-${sub}_task-${task}"
+		subDir="${derivDir}/${sub}/func"
+		tsv_prefix="${sub}_task-${task}"
 	fi
 	
 	# create events.tsv file for each subject who has functional data
 	if [ -d ${subDir} ] # if the subject has a functional data folder
 	then
-		echo "Generating ${task} events.tsv files for sub-${sub}"
+		echo "Generating ${task} events.tsv files for ${sub}"
 	
 		# if runs are specified (i.e., task includes several runs specified in file name)
 		if [[ ${runs} != 'NA' ]]
@@ -91,23 +81,32 @@ do
 			for r in $(echo ${runs} | tr "," "\n") # runs separated by commas in text file
 			do
 				# define event file for run
-				run_file="${tsv_prefix}_run-0${r}_events.tsv"
-			
+				run_file="${tsv_prefix}_run-00${r}_events.tsv"
+				
 				# delete events file if is already exists
 				if [ -f ${subDir}/${run_file} ]
 				then 
 					rm ${subDir}/${run_file}
 				fi	
 				
-				# copy the event files saved in the project files directory
-				for e in ${projDir}/files/event_files/${task}/*.tsv
-				do
-					awk 'NR>0' ${e} >> ${subDir}/tmp.tsv
-				done
+				# copy same onsets and durations for pixar or copy subject-specific events files
+				if [ ${study} == 'open-pixar' ]
+				then
+					# copy the event files saved in the project files directory
+					for e in ${projDir}/files/event_files/*.tsv
+					do
+						awk 'NR>0' ${e} >> ${subDir}/tmp.tsv
+					done
+					
+					# remove duplicate rows (header info) and tmp file
+					awk '!NF || !seen[$0]++' ${subDir}/tmp.tsv >> ${subDir}/${run_file}
+					rm ${subDir}/tmp.tsv
 				
-				# remove duplicate rows (header info) and tmp file
-				awk '!NF || !seen[$0]++' ${subDir}/tmp.tsv >> ${subDir}/${run_file}
-				rm ${subDir}/tmp.tsv
+				else
+					# copy the subject-specific event files saved in the project files directory
+					cp ${projDir}/files/event_files/${sub}/*run-00${r}*.tsv ${subDir}
+				fi
+
 			done
 		# if no run information is specified (i.e., only 1 run or not indicated in filenames)
 		else
@@ -119,18 +118,27 @@ do
 			then 
 				rm ${subDir}/${run_file}
 			fi	
+
+			# copy same onsets and durations for pixar or copy subject-specific events files
+			if [ ${study} == 'open-pixar' ]
+			then
+				# copy the event files saved in the project files directory
+				for e in ${projDir}/files/event_files/*.tsv
+				do
+					awk 'NR>0' ${e} >> ${subDir}/tmp.tsv
+				done
+				
+				# remove duplicate rows (header info) and tmp file
+				awk '!NF || !seen[$0]++' ${subDir}/tmp.tsv >> ${subDir}/${run_file}
+				rm ${subDir}/tmp.tsv
 			
-			# copy the event files saved in the data directory
-			for e in ${projDir}/files/event_files/${task}/*.tsv
-			do
-				awk 'NR>0' ${e} >> ${subDir}/tmp.tsv
-			done
-			
-			# remove duplicate rows (header info) and tmp file
-			awk '!NF || !seen[$0]++' ${subDir}/tmp.tsv >> ${subDir}/${run_file}
-			rm ${subDir}/tmp.tsv	
+			else
+				# copy the subject-specific event files saved in the project files directory
+				cp ${projDir}/files/event_files/${sub}/*run-00${r}*.tsv ${subDir}
+			fi
 		fi
 	fi
 	
 done <$1
+
